requirements.txt
Django==3.2
Pillow==8.2.0
hachoir==3.1.2
mysqlclient==2.0.3
channels==3.0.3
requests
